Add initial toggles
These work in realtime via a websocket call, and can be called via URL GET commands, but haven't implemented that in TD yet.

Currently the overlay stops showing up if you switch camera feeds - I think it's because the video goes in front of the canvas, but need to investigate more.