Initial swing at image segmentation
Currently it works, but extracting the results is tough. Getting the RGBA 8 bit mask takes ~30ms.

It looks like the beginnings of WebGL texture output are in MediaPipe that will make this much faster to render to a WebGL component (or better MediaPipe will do this for you), but currently the categoryMask.getAsWebGLTexture() function just returns an empty texture.

Will work on implementing the RGBA version anyway for now