Add check to see if webcam exists before throwing error
If you're loading this for the first time on a new machine, you'll be told to load a webcam that probably doesn't exist. Instead of throwing an error, carry on and let the page load so the webcam list can be updated. TD will then request a valid webcam and all is well. Allows for the error to still be thrown if the webcam does exist and it can't start.